Looking for build stability + closing #12, making it disabled by default as it is a big new experimental feature.

+++ splout-server/src/main/java/com/splout/db/dnode/HttpFileExchanger.java
@@ -229,7 +229,7 @@ public class HttpFileExchanger extends Thread implements HttpHandler {
 		} catch(Throwable t) {
 			log.error(t);
 			callback.onError(t, tablespace, partition, version, dest);
-			if(dest != null && dest.exists()) {
+			if(dest != null && dest.exists() && !t.getMessage().contains("Incoming file already being transferred")) {
 				dest.delete();
 			}
 		} finally {
